Make explanations have a text area, support blank lines in them
